+++ NEWS.md
@@ -1,4 +1,6 @@
 # flattabler (development version)
+* Add parameter `keep_col_names` to `unpivot` function.
+* Read the indicated sheets from a folder of Excel files in `read_excel_folder()`.
 
 # flattabler 2.0.0
 * New data model for `pivot_table` class, vectorial algorithm for `unpivot()`.